requirements.txt
pymilvus==2.4.0
pandas==2.2.2
numpy==1.26.4
flask==3.0.3
flask_cors==3.0.10
gevent==23.9.1
sentence-transformers==2.6.1
google-generativeai==0.5.0
